Move all the new hooks into report/hooks.
I'm not sure how this hsould be structured, but it certainly doesn't look helpful to have all the files in /interface, so at least having them in report which is the only place they are used is helpful.
