+++ src/interface/report/Results/index.tsx
@@ -15,13 +15,13 @@ import LoadingBar from 'interface/LoadingBar';
 import Panel from 'interface/Panel';
 import ReadableListing from 'interface/ReadableListing';
 import { RootState } from 'interface/reducers';
+import { Filter } from 'interface/report/hooks/useTimeEventFilter';
 import REPORT_HISTORY_TYPES from 'interface/REPORT_HISTORY_TYPES';
 import ResultsChangelogTab from 'interface/ResultsChangelogTab';
 import { getResultTab } from 'interface/selectors/url/report';
 import { hasPremium } from 'interface/selectors/user';
 import Tooltip from 'interface/Tooltip';
 import TooltipProvider from 'interface/TooltipProvider';
-import { Filter } from 'interface/useTimeEventFilter';
 import Config from 'parser/Config';
 import CharacterProfile from 'parser/core/CharacterProfile';
 import CombatLogParser from 'parser/core/CombatLogParser';
