+++ src/interface/report/index.tsx
@@ -2,13 +2,6 @@ import { wclGameVersionToExpansion } from 'game/VERSIONS';
 import ErrorBoundary from 'interface/ErrorBoundary';
 import makeAnalyzerUrl from 'interface/makeAnalyzerUrl';
 import NavigationBar from 'interface/NavigationBar';
-import useBossPhaseEvents from 'interface/useBossPhaseEvents';
-import useCharacterProfile from 'interface/useCharacterProfile';
-import useEventParser from 'interface/useEventParser';
-import useEvents from 'interface/useEvents';
-import useParser from 'interface/useParser';
-import usePhases, { SELECTION_ALL_PHASES } from 'interface/usePhases';
-import useTimeEventFilter, { Filter } from 'interface/useTimeEventFilter';
 import Config from 'parser/Config';
 import { CombatantInfoEvent } from 'parser/core/Events';
 import { WCLFight } from 'parser/core/Fight';
@@ -21,6 +14,13 @@ import BOSS_PHASES_STATE from './BOSS_PHASES_STATE';
 import ConfigContext from './ConfigContext';
 import EVENT_PARSING_STATE from './EVENT_PARSING_STATE';
 import FightSelection from './FightSelection';
+import useBossPhaseEvents from './hooks/useBossPhaseEvents';
+import useCharacterProfile from './hooks/useCharacterProfile';
+import useEventParser from './hooks/useEventParser';
+import useEvents from './hooks/useEvents';
+import useParser from './hooks/useParser';
+import usePhases, { SELECTION_ALL_PHASES } from './hooks/usePhases';
+import useTimeEventFilter, { Filter } from './hooks/useTimeEventFilter';
 import PatchChecker from './PatchChecker';
 import PlayerLoader from './PlayerLoader';
 import ReportLoader from './ReportLoader';
